Fix SlashCommands import cycles
