+++ apps/web/src/slash-commands/utils.ts
@@ -76,3 +76,23 @@ export const isCurrentLocalRoom = (cli: MatrixClient | null): boolean => {
     if (!room) return false;
     return isLocalRoom(room);
 };
+
+/**
+ * Split the supplied string into one or two strings separated by the first
+ * region of white space we can find.
+ */
+export function splitAtFirstSpace(args: string): [string, string?] {
+    const trimmedArgs = args.trim();
+    const i = trimmedArgs.search(/\s+/);
+    if (i === -1) {
+        return [trimmedArgs];
+    } else {
+        const first = trimmedArgs.slice(0, i);
+        const second = trimmedArgs.slice(i + 1).trimStart();
+        if (second === "") {
+            return [first];
+        } else {
+            return [first, second];
+        }
+    }
+}
